+++ src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Commands/Create/CreateDialogCommandValidator.cs
@@ -2,10 +2,9 @@
 using Digdir.Domain.Dialogporten.Application.Common.Extensions.Enumerables;
 using Digdir.Domain.Dialogporten.Application.Common.Extensions.FluentValidation;
 using Digdir.Domain.Dialogporten.Application.Externals.Presentation;
-using Digdir.Domain.Dialogporten.Application.Features.V1.Common.Actors;
 using Digdir.Domain.Dialogporten.Application.Features.V1.Common.Content;
 using Digdir.Domain.Dialogporten.Application.Features.V1.Common.Localizations;
-using Digdir.Domain.Dialogporten.Domain.Actors;
+using Digdir.Domain.Dialogporten.Application.Features.V1.ServiceOwner.Common.Actors;
 using Digdir.Domain.Dialogporten.Domain.Common;
 using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Actions;
 using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Activities;
@@ -150,7 +149,7 @@ internal sealed class CreateDialogCommandValidator : AbstractValidator<CreateDia
 internal sealed class CreateDialogDialogTransmissionDtoValidator : AbstractValidator<TransmissionDto>
 {
     public CreateDialogDialogTransmissionDtoValidator(
-        IValidator<TransmissionSenderActorDto> actorValidator,
+        IValidator<ActorDto> actorValidator,
         IValidator<TransmissionContentDto> contentValidator,
         IValidator<TransmissionAttachmentDto> attachmentValidator)
     {
@@ -409,7 +408,7 @@ internal sealed class CreateDialogDialogActivityDtoValidator : AbstractValidator
 {
     public CreateDialogDialogActivityDtoValidator(
         IValidator<IEnumerable<LocalizationDto>> localizationsValidator,
-        IValidator<ActivityPerformedByActorDto> actorValidator)
+        IValidator<ActorDto> actorValidator)
     {
         RuleFor(x => x.Id)
             .IsValidUuidV7()
@@ -441,43 +440,5 @@ internal sealed class CreateDialogDialogActivityDtoValidator : AbstractValidator
             .NotEmpty()
             .WithMessage($"A {nameof(DialogActivityType.Values.TransmissionOpened)} needs to reference a transmission.")
             .When(x => x.Type == DialogActivityType.Values.TransmissionOpened);
-    }
-}
-
-internal sealed class CreateDialogDialogTransmissionActorDtoValidator : AbstractValidator<TransmissionSenderActorDto>
-{
-    public CreateDialogDialogTransmissionActorDtoValidator()
-    {
-        RuleFor(x => x.ActorType)
-            .IsInEnum();
-
-        RuleFor(x => x)
-            .Must(dto => (dto.ActorId is null || dto.ActorName is null) &&
-                         ((dto.ActorType == ActorType.Values.ServiceOwner && dto.ActorId is null && dto.ActorName is null) ||
-                          (dto.ActorType != ActorType.Values.ServiceOwner && (dto.ActorId is not null || dto.ActorName is not null))))
-            .WithMessage(ActorValidationErrorMessages.ActorIdActorNameExclusiveOr);
-
-        RuleFor(x => x.ActorId!)
-            .IsValidPartyIdentifier()
-            .When(x => x.ActorId is not null);
-    }
-}
-
-internal sealed class CreateDialogDialogActivityActorDtoValidator : AbstractValidator<ActivityPerformedByActorDto>
-{
-    public CreateDialogDialogActivityActorDtoValidator()
-    {
-        RuleFor(x => x.ActorType)
-            .IsInEnum();
-
-        RuleFor(x => x)
-            .Must(dto => (dto.ActorId is null || dto.ActorName is null) &&
-                         ((dto.ActorType == ActorType.Values.ServiceOwner && dto.ActorId is null && dto.ActorName is null) ||
-                          (dto.ActorType != ActorType.Values.ServiceOwner && (dto.ActorId is not null || dto.ActorName is not null))))
-            .WithMessage(ActorValidationErrorMessages.ActorIdActorNameExclusiveOr);
-
-        RuleFor(x => x.ActorId!)
-            .IsValidPartyIdentifier()
-            .When(x => x.ActorId is not null);
     }
 }