+++ src/Digdir.Domain.Dialogporten.WebApi/Common/Constants.cs
@@ -5,6 +5,7 @@ internal static class Constants
     internal const string IfMatch = "If-Match";
     internal const string Authorization = "Authorization";
     internal const string CurrentTokenIssuer = "CurrentIssuer";
+    internal const int MaxRequestBodySize = 100_000;
 
     internal static class SwaggerSummary
     {